pollution_v2/dispersal/requirements.txt
numpy==1.22.0
pandas==1.3.5
PyYAML==6.0.2
fastapi[standard]
GDAL==3.4.3
python-dotenv==1.0.1
pytz==2023.3
rasterio==1.3.10
